refactor: extract compile runtime helpers into compile_runtime module
